[LinuxOS] Backports regarding Linux::SetSystemCallStatus modifications; Linux syscall return status is signed long, thus generic SetSystemCallStatus has been update to read int64_t values.
Former-commit-id: ca3a3935c2ac6c483789e0582a051b35235d6e93

+++ unisim_lib/unisim/util/os/linux_os/arm.hh
@@ -277,7 +277,7 @@ namespace linux_os {
     
     static void SetARMSystemCallStatus(LINUX& _lin, int ret, bool error) { SetRegister(_lin, kARM_r0, (parameter_type) ret); }
     
-    void SetSystemCallStatus(int ret, bool error) const { SetARMSystemCallStatus( lin, ret, error ); }
+    void SetSystemCallStatus(int64_t ret, bool error) const { SetARMSystemCallStatus( lin, ret, error ); }
     
     static parameter_type GetSystemCallParam( LINUX& _lin, int id )
     {
